requirements.txt
requests>=2.31.0
colorama>=0.4.6
tabulate>=0.9.0
beautifulsoup4>=4.12.0
urllib3>=2.0.0
python-Wappalyzer
pyppeteer
selenium>=4.10.0
chromedriver-autoinstaller>=0.6.0
pyOpenSSL>=23.2.0
